+++ docs/development/changelog.md
@@ -2,6 +2,50 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### 0.0.9 (2024-05-01)
+
+
+### 🐛 Bug Fixes | Bug 修复
+
+* 解决触摸设备下事件报错问题，增加默认样式绑定及一些manager options ([944155c](https://gitee.com/sunhy-cn/touch-fusion/commit/944155ccc28e6ca9ea6b0560c8282d07bb4afd48))
+
+
+### ✅ Tests | 测试
+
+* 修改一些单元测试文件内容引用错误 ([842a0b1](https://gitee.com/sunhy-cn/touch-fusion/commit/842a0b156d8550f7dc96ffeb545117e020b33cd6))
+
+
+### ✏️ Documentation | 文档
+
+* [style]修改API Table颜色变量 ([e47abde](https://gitee.com/sunhy-cn/touch-fusion/commit/e47abde874f5d0bc8a87ff7b2aee590b9b3ae616))
+* 优化长按手势下右键事件禁用区域 ([d905fdf](https://gitee.com/sunhy-cn/touch-fusion/commit/d905fdf77ccdf7861d0b6ba39e659a4837a0676a))
+* 修改git仓库地址链接 ([2156571](https://gitee.com/sunhy-cn/touch-fusion/commit/21565711ff11f305acb843534b58f2658e1663e4))
+* 修改文档页根目录 ([42ca57c](https://gitee.com/sunhy-cn/touch-fusion/commit/42ca57c7172fb85db9bc85d6a31a8b8ac4df632a))
+* 修改项目导出结构 ([15f18c3](https://gitee.com/sunhy-cn/touch-fusion/commit/15f18c30eceb865e80429864498431d715c2b206))
+* 增加Changelog文件 ([51ae827](https://gitee.com/sunhy-cn/touch-fusion/commit/51ae82791dd0bc139f22568a3c9827cb04523da0))
+* 增加使用文档 ([2132e13](https://gitee.com/sunhy-cn/touch-fusion/commit/2132e13dc62984232017f1242db2e68250ba5df1))
+* 增加多指长按手势 ([b582b7e](https://gitee.com/sunhy-cn/touch-fusion/commit/b582b7e3e95eeb08ba5e26924cde927ef72ecff0))
+* 增加拖拽手势文档 ([684190f](https://gitee.com/sunhy-cn/touch-fusion/commit/684190fad031e177c3fe1ca12dd555c9bf60b64f))
+* 增加文档演示组件 ([8efb781](https://gitee.com/sunhy-cn/touch-fusion/commit/8efb7812cc2fff3b7398a1abd7188d889429f135))
+* 增加文档目录 ([b98181b](https://gitee.com/sunhy-cn/touch-fusion/commit/b98181bbddd36d6237c52bfd4f9f64fafc1e3a13))
+* 增加滑动手势文档 ([a52a5e5](https://gitee.com/sunhy-cn/touch-fusion/commit/a52a5e5bf987c424e6ab9b8601ba2654b67b2da7))
+* 更新文档构建内容 ([adeaf15](https://gitee.com/sunhy-cn/touch-fusion/commit/adeaf158d29ae77942e8de5c80cef86bbdc0d5db))
+* 更新长按手势文档 ([74ae526](https://gitee.com/sunhy-cn/touch-fusion/commit/74ae526c31bf27ae895d8cb7b23eb1d4c048fd6b))
+
+
+### ✨ Features | 新功能
+
+* 增加PointerEvent的支持 ([5d29427](https://gitee.com/sunhy-cn/touch-fusion/commit/5d29427105142b8d764b6f247de3976d23ad8144))
+* 增加滑动手势 ([9a833fd](https://gitee.com/sunhy-cn/touch-fusion/commit/9a833fda1aec1c3c84483fd72d4f483d56347275))
+* 完善Manager类 ([1b253af](https://gitee.com/sunhy-cn/touch-fusion/commit/1b253af32b0fe97c8f25ed04c396ce35e6586a58))
+* 完成拖拽手势 ([85c6c70](https://gitee.com/sunhy-cn/touch-fusion/commit/85c6c70dea1f1c6042281651da28d13a59c85c01))
+* 完成捏合/开合手势识别 ([a0fd4b0](https://gitee.com/sunhy-cn/touch-fusion/commit/a0fd4b0e615716487f64a03bb59aa2cce42f4ce2))
+* 完成输入事件统一化处理 ([f037a3e](https://gitee.com/sunhy-cn/touch-fusion/commit/f037a3e471a644232eca90a84280855609c4359a))
+* 实现事件输入基类 ([99b2b08](https://gitee.com/sunhy-cn/touch-fusion/commit/99b2b082eceadbfae2625c3624d6fc161abfe04f))
+* 实现识别器基类 ([5de75a7](https://gitee.com/sunhy-cn/touch-fusion/commit/5de75a7cd90e9d1e3cbd6a36afb82a3e2dcb9952))
+* 实现长按事件识别器 ([c59fb72](https://gitee.com/sunhy-cn/touch-fusion/commit/c59fb72d4af7ef9f873a1570a9a8188363cc08cc))
+* 支持长按手势 ([ce58cbd](https://gitee.com/sunhy-cn/touch-fusion/commit/ce58cbd248c13eed88a35bb6d65829e0011b3e0f))
+
 ### 0.0.8 (2024-04-15)